Improve concurrent file access
If the device file is opened twice concurrently, closing one of them
will clear configurations such as the timer and fast_buf, leading to
errors in the other operation. Therefore, considering the state under
concurrency, an atomic_t type variable named open_cnt is added to
track the number of opens. The original settings operations will only
be performed when it is the first open or the last close.