Ensure v6.4 compatibility (#1)
In the previous implementation in the function simrupt_init(), we
utilize a function call "class_create()" to create a class structure for
the kernel module. This function is explicitly called with
"class_create(THIS_MODULE, DEV_NAME)", which will cause a compiling
error. The reason is the Linux Kernel removes the module pointer from
the function "class_create()" in v6.4.0 and later versions. We should
support the capability for versions above v6.4.0 as the corresponding
changes have made.